test-1/requirements.txt
streamlit==1.22.0
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib>=2.0.0
pytest==7.3.1
azure-cognitiveservices-speech==1.43.0
pydub==0.25.1
openai>=0.27.0
google-api-python-client>=2.118.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
streamlit-calendar==1.3.1